Sample Card Game
The following application is a real time web based card game with a
similar concept to Top Trumps playing cards. The application allows a
user to log in (a username can only be used by one user, this is
handled by the server app) and challenge other users logged in to a
card game. When a user excepts a challenge a game object will be
created on the server side and the game will start on both client web
browsers. The game will allocate 15 cards to each user and each user
will have consecutive turns to pick an attribute on their card with the
corresponding value on the other client's card. All of this logic
happens on the server side. To put the game in its simplest form the
client and server send text data to each other and events are triggered
on the based of the data.